+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Next Release
 
+## 0.0.74
+
+- [sdk] Integrate Gelato for fulfill relaying
+- [contracts] Include `remove-router` hardhat task
 - [sdk] SDK supports string `subgraph` in config
 
 ## 0.0.73